chore(ai-reviews): move sendReviewNotification task into ee/
